Update of gmp++ (shift operators, long long portable capabilities) Removal of linbox-config.h as weel as givaro-config.h for portabiblity Definition of GMP_VERSION_3 macro instead of GMP_VERSION_4 for future genericity

+++ trunk/linbox/linbox/util/gmp++/gmp++_int_io.C
@@ -65,7 +65,7 @@ Integer::Integer(const std::vector<mp_limb_t>& v) {
 
 		std::vector<mp_limb_t>::const_iterator vi = v.begin();
 		for(++vi;vi != v.end();++vi) { 
-	 		*this += ( prod * ((unsigned long) (*vi)) );
+	 		*this += ( prod * (*vi) );
 			prod *= base;
 		}
 	} else
